Simplify and cleanup yaml IO

+++ gammapy/utils/scripts.py
@@ -8,17 +8,21 @@ from pathlib import Path
 from uuid import uuid4
 import yaml
 from gammapy.utils.check import add_checksum, verify_checksum
+from gammapy.utils.deprecation import deprecated_renamed_argument
 
 __all__ = [
+    "from_yaml",
     "get_images_paths",
     "make_path",
     "read_yaml",
     "recursive_merge_dicts",
+    "to_yaml",
     "write_yaml",
 ]
 
 PATH_DOCS = Path(__file__).resolve().parent / ".." / ".." / "docs"
 SKIP = ["_static", "_build", "_checkpoints", "docs/user-guide/model-gallery/"]
+YAML_FORMAT = dict(sort_keys=False, indent=4, width=80, default_flow_style=False)
 
 
 def get_images_paths(folder=PATH_DOCS):
@@ -32,6 +36,36 @@ def get_images_paths(folder=PATH_DOCS):
     for i in Path(folder).rglob("images/*"):
         if not any(s in str(i) for s in SKIP):
             yield i.resolve()
+
+
+def from_yaml(text, sort_keys=False, checksum=False):
+    """Read YAML file.
+
+    Parameters
+    ----------
+    text : str
+        yaml str
+    sort_keys : bool, optional
+        Whether to sort keys. Default is False.
+    checksum : bool
+        Whether to perform checksum verification. Default is False.
+
+    Returns
+    -------
+    data : dict
+        YAML file content as a dictionary.
+
+    """
+    data = yaml.safe_load(text)
+    checksum_str = data.pop("checksum", None)
+    if checksum:
+        yaml_format = YAML_FORMAT.copy()
+        yaml_format["sort_keys"] = sort_keys
+        yaml_str = yaml.dump(data, **yaml_format)
+        if not verify_checksum(yaml_str, checksum_str):
+            warnings.warn("Checksum verification failed.", UserWarning)
+
+    return data
 
 
 def read_yaml(filename, logger=None, checksum=False):
@@ -56,24 +90,38 @@ def read_yaml(filename, logger=None, checksum=False):
         logger.info(f"Reading {path}")
 
     text = path.read_text()
-
-    data = yaml.safe_load(text)
-    checksum_str = data.pop("checksum", None)
-    if checksum:
-        index = text.find("checksum")
-        if not verify_checksum(text[:index], checksum_str):
-            warnings.warn(f"Checksum verification failed for {filename}.", UserWarning)
-
-    return data
+    return from_yaml(text, checksum=checksum)
 
 
-def write_yaml(dictionary, filename, logger=None, sort_keys=True, checksum=False):
-    """Write YAML file.
+def to_yaml(dictionary, sort_keys=False):
+    """dict to yaml
 
     Parameters
     ----------
     dictionary : dict
         Python dictionary.
+    sort_keys : bool, optional
+        Whether to sort keys. Default is False.
+    """
+    from gammapy.utils.metadata import CreatorMetaData
+
+    yaml_format = YAML_FORMAT.copy()
+    yaml_format["sort_keys"] = sort_keys
+    text = yaml.safe_dump(dictionary, **yaml_format)
+    creation = CreatorMetaData()
+    return text + creation.to_yaml()
+
+
+@deprecated_renamed_argument("dictionary", "text", "v1.3")
+def write_yaml(
+    text, filename, logger=None, sort_keys=False, checksum=False, overwrite=False
+):
+    """Write YAML file.
+
+    Parameters
+    ----------
+    text : str
+        yaml str
     filename : `~pathlib.Path`
         Filename.
     logger : `~logging.Logger`, optional
@@ -82,14 +130,17 @@ def write_yaml(dictionary, filename, logger=None, sort_keys=True, checksum=False
         Whether to sort keys. Default is True.
     checksum : bool, optional
         Whether to add checksum keyword. Default is False.
+    overwrite : bool, optional
+        Overwrite existing file. Default is False.
     """
-    text = yaml.safe_dump(dictionary, default_flow_style=False, sort_keys=sort_keys)
 
     if checksum:
         text = add_checksum(text, sort_keys=sort_keys)
 
     path = make_path(filename)
     path.parent.mkdir(exist_ok=True)
+    if path.exists() and not overwrite:
+        raise IOError(f"File exists already: {path}")
     if logger is not None:
         logger.info(f"Writing {path}")
     path.write_text(text)
